Revert "arm64: lib: memset: add a missing reference to Lzero_mem"
This reverts commit cd95e345f1751da40103085d3a89dfd16ab40f75, as it is
the one that causes an issue with audio not working after booting in and
out of recovery, unless the recovery also has a kernel with that commit.
In practice, I guess it is a little inconvenient with usage, by others,
of other older recovery builds built from other kernel sources without
it.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>